Fix and enhance new mastodon profile links
This is a follow-up to [cd19a17b].
The fontawesome sources used by the website are too old to have captured the
mastodon icon. As a simple workaround, this changeset adds the icon's
svg. In addition, it adds the rel="me" attribute to the mastodon links,
which allows team members to verify their mastodon accounts through
these links.